Add rm() function to remove files and directories
This commit adds a new function `rm()` to the `FileSystem` class. The `rm()` function allows for the removal of files and directories from the file system. It checks if the given path exists and then determines whether it is a file or a directory. If it is a directory, the function calls the `rmdir()` function to remove the directory. If it is a file, the function calls the `rmfile()` function to remove the file. This new functionality enhances the capabilities of the file system.