Add recursive file removal function
This commit adds a new member function `rm_recursive` to the `FileSystem` class. The function allows for the recursive removal of files and directories. It checks if the specified path exists and throws an exception if it doesn't. If the path is a file, it is removed using the `rmfile` function. If the path is a directory, it recursively removes all files and directories within it before removing the directory itself using the `rmdir` function. Exceptions are thrown if the root directory or the working directory is attempted to be removed.